Add two GM crucibles and macro to roll them
This commit adds compendia to the system - one called Crucibles, which
can be used for storing crucibles as rolltables and one called
Grimwild Macros, which can be used for storing macros.

Currently there are the two GM crucibles loaded into the Crucibles.
These are implemented as 1d36 rolltables which have each possible 2d6
outcome coded in as a pair of options.

The macro simply rolls both tables and presents the options conveniently
for the user.