Slight improve roll dialog logic
This commit moves the computation of ignoring marks into the roll dialog
app as the actor itself should not be concerned with the logic of
whether to show that or not.

This commit also ignores assists with no dice, and adds a default name
of "Assist" to assists with no name so the chat message doesn't have an
arrow pointing to nothing.