Add Monsters compendium
This commit adds a new system compendium with all the monsters from
Chapter 5 of the rulebook. Credit for this incredible work goes to One
Proud Bavarian (oneproudbavarian) from the Oddity Press discord server
who took the time to do the data entry for all the monsters. I simply
did the work of converting it into a system compendium.

Notes from oneproudbavarian:
* I left all the monsters on the automatically used tier "Tough" as
  Chapter 5 doesn't firmly assign anything to them and this is up to the
GMs.
* I included a dragon actor that includes all Breath types and then
  individual Dragon actor in the "Dragon,[Type]" format that only
include one specific breath type. I did a similar thing for Ettin -
there is an Ettin actor with all the Ettin heads inside of the sheet and
then three Ettin actors in the format "Ettin,[Head 1] & [Head 2]" that
match the addigned PDF heads.
* I included any linked challenges present in Chapter 5, which means the
  Giants for example have their, but the Great Red Dragon example linked
challenge is not included.
* The big weakness currently definitely rests with linked challenges...
  I think most linked challenges are legible for anybody that looks at
them for the first time (Demon, Balor and Hydra for example should
intuitively make sense), but the Ettercap for example is incredibly
complex and isn't really interpretable for GMs without the numerous
connections being displayed.